GH-35: Move Element into core
